[Misc] Move shop lateral menu mixin into shops. Create helpers mixins
