+++ borgia/mixins.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.mixins import PermissionRequiredMixin
 from django.contrib.auth.models import Group
 from django.urls import reverse
 from django.views.generic.base import ContextMixin
@@ -249,163 +250,13 @@ class LateralMenuManagersMixin(LateralMenuBaseMixin):
         return nav_tree
 
 
-class LateralMenuShopsMixin(LateralMenuBaseMixin):
+class MembersMixin(PermissionRequiredMixin, LateralMenuMembersMixin):
     """
-    Lateral Menu for shops managers.
-
-    Add :
-    - Home page shops
-    - Checkup
-    - OperatorSale module
-    - Sales
-    - Products
-    - StockEntries
-    - Inventories
-    - OperatorSale Configuration
-    - SelfSale Configuration
-    - Shop groups management
+    Mixin that check permission and add MEMBERS lateral menu.
     """
 
-    def lateral_menu(self):
-        nav_tree = super().lateral_menu()
-        user = self.request.user
-        shop = self.shop
 
-        nav_tree.append(
-            simple_lateral_link(
-                'Accueil Magasin ' + shop.name.title(),
-                'briefcase',
-                'lm_workboard',
-                reverse('url_group_workboard')))
-
-        if user.has_perm('shops.view_shop'):
-            nav_tree.append(
-                simple_lateral_link(
-                    'Checkup',
-                    'user',
-                    'lm_shop_checkup',
-                    reverse('url_shop_checkup'))
-            )
-
-        # OperatorSale Module
-        if shop.modules_operatorsalemodule_shop.first().state is True:
-            if user.has_perm('modules.use_operatorsalemodule'):
-                nav_tree.append(simple_lateral_link(
-                    label='Module vente',
-                    fa_icon='shopping-basket',
-                    id_link='lm_operatorsale_interface_module',
-                    url=reverse(
-                        'url_shop_module_sale',
-                        kwargs={'shop_pk': shop.pk, 'module_class': 'operator_sales'})
-                    )
-                )
-
-        # Sales
-        if user.has_perm('finances.view_sale'):
-            nav_tree.append(
-                simple_lateral_link(
-                    'Ventes',
-                    'shopping-cart',
-                    'lm_sale_list',
-                    reverse(
-                        'url_sale_list', 
-                        kwargs={'shop_pk': shop.pk})
-                )
-            )
-
-        # Products
-        if user.has_perm('shops.view_product'):
-            nav_tree.append(
-                simple_lateral_link(
-                    label='Produits',
-                    fa_icon='cube',
-                    id_link='lm_product_list',
-                    url=reverse(
-                        'url_product_list',
-                        kwargs={'shop_pk': shop.pk})
-                )
-            )
-
-        # StockEntries
-        if user.has_perm('stocks.view_stockentry'):
-            nav_tree.append(
-                simple_lateral_link(
-                    'Entrées de stock',
-                    'list',
-                    'lm_stockentry_list',
-                    reverse(
-                        'url_stockentry_list', 
-                        kwargs={'shop_pk': shop.pk})
-                )
-            )
-        
-        # Inventories
-        if user.has_perm('stocks.view_stockentry'):
-            nav_tree.append(
-                simple_lateral_link(
-                    'Entrées de stock',
-                    'list',
-                    'lm_inventory_list',
-                    reverse(
-                        'url_inventory_list', 
-                        kwargs={'shop_pk': shop.pk})
-                )
-            )
-
-        if user.has_perm('modules.view_config_selfsalemodule'):
-            nav_tree.append(
-                simple_lateral_link(
-                    label='Configuration vente libre service',
-                    fa_icon='shopping-basket',
-                    id_link='lm_selfsale_module',
-                    url=reverse('url_shop_module_config',
-                                kwargs={'shop_pk': shop.pk, 'module_class': 'self_sales'}
-                                )
-                ))
-
-        if user.has_perm('modules.view_config_operatorsalemodule'):
-            nav_tree.append(
-                simple_lateral_link(
-                    label='Configuration vente par opérateur',
-                    fa_icon='coffee',
-                    id_link='lm_operatorsale_module',
-                    url=reverse('url_shop_module_config',
-                                kwargs={'shop_pk': shop.pk, 'module_class': 'operator_sales'}
-                                )
-                ))
-
-        # Groups management
-        nav_management_groups = {
-            'label': 'Gestion groupes magasin',
-            'icon': 'users',
-            'id': 'lm_group_management',
-            'subs': []
-        }
-        groups = [Group.objects.get(name='chiefs-'+shop.name), Group.objects.get(name='associates-'+shop.name)]
-        for group in groups:
-            if user.has_perm(get_permission_name_group_managing(group)):
-                nav_management_groups['subs'].append(
-                        simple_lateral_link(
-                        'Gestion ' + group_name_display(group),
-                        'users',
-                        'lm_group_manage_' + group.name,
-                        reverse('url_group_update', kwargs={
-                            'pk': group.pk})
-                    ))
-
-        nav_tree.append(nav_management_groups)
-
-
-
-        if self.lm_active is not None:
-            for link in nav_tree:
-                try:
-                    for sub in link['subs']:
-                        if sub['id'] == self.lm_active:
-                            sub['active'] = True
-                            break
-                except KeyError:
-                    if link['id'] == self.lm_active:
-                        link['active'] = True
-                        break
-        return nav_tree
+class ManagerMixin(PermissionRequiredMixin, LateralMenuManagersMixin):
+    """
+    Mixin that check permission and add MANAGERS lateral menu.
+    """
